soundQuize/requirements.txt
requests>=2.31.0
tqdm>=4.66.0
playsound3==3.3.0
huggingface-hub>=0.20.0
